NC-45 Big Navbar cut into 3 pieces
